requirements.txt
bcrypt==4.2.0
pycryptodome==3.21.0
tabulate==0.9.0

